+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.2.0 (2021-02-08)
+
+
+### Features
+
+* :sparkles: added connection search. ([6b44e98](https://gitlab.com/stream-app-io/knecthub-client/-/commit/6b44e987dd0b95b82f49ae20115e60aebf987df1))
+* :sparkles: added profile update. ([763f9cc](https://gitlab.com/stream-app-io/knecthub-client/-/commit/763f9cc0668e0aa3c6794fcdae7efa5bbe9a8d92))
+* :tada: initial connection. ([89972c5](https://gitlab.com/stream-app-io/knecthub-client/-/commit/89972c5dae4035c0a91dbde319768a78f4808b6d))
+* **response:** :sparkles: added standard response. ([74ce4b2](https://gitlab.com/stream-app-io/knecthub-client/-/commit/74ce4b2705d29b6188dcd1a1050e7d89c62c0bb6))
+* **User:** :sparkles: supports multiple emails. ([9ec8453](https://gitlab.com/stream-app-io/knecthub-client/-/commit/9ec84538d9c84065ae8771128577ac1b8813a7b2))
+* **User:** :sparkles: update primary email. ([3d0c0ee](https://gitlab.com/stream-app-io/knecthub-client/-/commit/3d0c0ee7e46515d20223faf9b333ba15068eb747))
+* :sparkles: added fallback response. ([879cca9](https://gitlab.com/stream-app-io/knecthub-client/-/commit/879cca9979751b776ad655dfc0d3a874b493ef15))
+
+
+### Bug Fixes
+
+* :bug: deprecated url format. ([919717c](https://gitlab.com/stream-app-io/knecthub-client/-/commit/919717c731406bc18004f532822d5ce3ef52ebe4))
+* :bug: primary email confirmed property not updating. ([a811d72](https://gitlab.com/stream-app-io/knecthub-client/-/commit/a811d724587e1a40f4da1a721b011fded86d2018))
+* :bug: remove id when _id is null. ([d27b6f4](https://gitlab.com/stream-app-io/knecthub-client/-/commit/d27b6f4bfb0c912a948a8331224a188f3f218054))
+* :bug: update response to StandardResponse. ([b9d1bfb](https://gitlab.com/stream-app-io/knecthub-client/-/commit/b9d1bfbe9ed5eb03243deffa9989a21c341e783f))
+* **Registration:** :bug: verification link email query params. ([cda517f](https://gitlab.com/stream-app-io/knecthub-client/-/commit/cda517f5474cab0fecb12461ee575fd8f2d8d459))
+* **response:** :bug: change error response from array to single object. ([d827687](https://gitlab.com/stream-app-io/knecthub-client/-/commit/d8276876488a880d6fc8566dba8ca89ff39b1de4))
+* **response:** :poop: return the updated user after confirming email. ([483e84f](https://gitlab.com/stream-app-io/knecthub-client/-/commit/483e84f22e47382ed8f74beb9207f074cf877553))
+
 ## 1.1.0 (2021-02-05)